pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
# name it as your package name
name = "quanttide-doc-agents"
# semetric versions
version = "0.1.0-alpha.1"
# describe the package within one sentence
description = "QuantTide Documentation Agents"
authors = [
    {name = "QuantTide Inc.", email = "opensource@quanttide.com"},
    {name = "Guo Zhang", email = "zhangguo@quanttide.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = '>=3.8'
dependencies = [
    "dashscope>=1.19.2",
    "python-dotenv>=1.0.1",
]
# dynamic = ["version"]

[project.license]
# create your license and name it as "LICENSE"
file = "LICENSE"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[tool.pdm.build]
includes = [
    "quanttide_doc_agents",
]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.2",
    "coverage>=7.5.3",
]
